organize src folder
create Math subfolder for mathematics files
