+++ src/Coding/DecodeSyndrome.py
@@ -1,5 +1,5 @@
-from src.FiniteField import CARDINAL, MESSAGE_SIZE, F, DEGREE_POLYNOME_Q0, DEGREE_POLYNOME_Q1, power, X
-from src.Matrix import CONTROL_MATRIX, ENCODE_MATRIX_INVERSE, syndrome_extracted, control_extracted
+from src.Math.FiniteField import CARDINAL, MESSAGE_SIZE, F, DEGREE_POLYNOME_Q0, DEGREE_POLYNOME_Q1, power, X
+from src.Math.Matrix import CONTROL_MATRIX, ENCODE_MATRIX_INVERSE, syndrome_extracted, control_extracted
 
 
 def decode_syndrome(message: list)->list:
