+++ src/Coding/Encode.py
@@ -1,4 +1,4 @@
-from src.Matrix import ENCODE_MATRIX
+from src.Math.Matrix import ENCODE_MATRIX
 
 def encode(message):
     return ENCODE_MATRIX.LeftMulColumnVec(message)
